requirements.txt
aiofiles==24.1.0
aiohttp==3.11.18
aiosqlite==0.21.0
beautifulsoup4==4.13.4
chromadb>=1.0.8
deep_translator==1.11.4
discord-pretty-help
discord.py==2.5.2
discord.py[voice]
duckduckgo_search==7.5.2
fake_useragent==2.1.0
fastapi==0.115.9
huggingface_hub==0.30.2
jinja2==3.1.6
ollama==0.4.8
openai==1.77.0
orjson==3.10.18
parsedatetime==2.6
Pillow==11.2.1
python-dotenv==1.1.0
python-multipart==0.0.20
pytubefix==8.12.0
qrcode==8.2
Requests==2.32.3
scrapetube==2.5.1
tenacity==9.0.0
tweepy==4.15.0
uvicorn==0.34.2
yt_dlp
zhipuai==2.1.5.20250421